Fix Event update Add ClubOkResponse class and ClubOkSuccessResponse, ClubOkErrorResponse subclasses Change ClubOkException Change Validation error handling Add Administration route

+++ clubok_be/src/main/java/dc/clubok/db/controllers/EventController.java
@@ -1,7 +1,7 @@
 package dc.clubok.db.controllers;
 
 import dc.clubok.db.models.Event;
-import dc.clubok.utils.exceptions.ClubOkException;
+import dc.clubok.utils.ClubOkException;
 import org.bson.Document;
 import org.bson.conversions.Bson;
 import org.slf4j.Logger;
@@ -9,7 +9,9 @@ import org.slf4j.LoggerFactory;
 
 import java.util.List;
 
+import static dc.clubok.utils.Constants.ERROR_QUERY;
 import static dc.clubok.utils.Constants.model;
+import static org.apache.http.HttpStatus.SC_NOT_FOUND;
 
 public class EventController {
 
@@ -39,10 +41,22 @@ public class EventController {
 //    }
 
     public static void deleteEventById(String eventId) throws ClubOkException {
+        Event event = getEventById(eventId);
+        if (event == null) {
+            Document details = new Document("details", "Such event does not exist");
+            throw new ClubOkException(ERROR_QUERY, details, SC_NOT_FOUND);
+        }
+
         model.deleteById(eventId, Event.class);
     }
 
-    public static void editEventById(String eventId, Event event) throws ClubOkException {
-        model.modify(event, new Document("_id", eventId), Event.class);
+    public static void editEventById(String eventId, Document update) throws ClubOkException {
+        Event event = getEventById(eventId);
+        if (event == null) {
+            Document details = new Document("details", "Such event does not exist");
+            throw new ClubOkException(ERROR_QUERY, details, SC_NOT_FOUND);
+        }
+
+        model.modify(event, update, Event.class);
     }
 }
